Move NavLink to components
