+++ src/theme/NavbarItem/MegaDropdownNavbarItem.tsx
@@ -14,7 +14,7 @@ import {
   Collapsible,
   useLocalPathname,
 } from '@docusaurus/theme-common';
-import { NavLink } from './DefaultNavbarItem';
+import NavLink from '@site/src/components/NavLink';
 import NavbarItem from '@theme/NavbarItem';
 import './styles.css';
 import type { Props as NavLinkProps } from '@theme/NavbarItem/DefaultNavbarItem';
